requirements.txt
streamlit==1.17.0
Pillow==8.3.1
numpy==1.21.1
scikit-learn==0.24.2
tensorflow==2.7.0
